+++ building-material/dev.R
@@ -3,7 +3,7 @@ usethis::use_testthat()
 usethis::use_test("flux_fitting_quadratic")
 
 #to add used packages in the description
-usethis::use_package("")
+usethis::use_package("lifecycle")
 
 # to make the webpage
 usethis::use_pkgdown()
@@ -36,6 +36,9 @@ lintr::lint_package()
 
 # adding check action
 usethis::use_github_action(name = "check-standard", badge = TRUE)
+
+# deprecating stuff
+usethis::use_lifecycle()
 
 # add code coverage badge
 # rcompendium::add_github_actions_codecov()